+++ src/gamestates/cWinLoseState.h
@@ -2,7 +2,7 @@
 
 #include "cGameState.h"
 #include "controls/cKeyboardEvent.h"
-#include "sMouseEvent.h"
+#include "controls/sMouseEvent.h"
 
 class Texture;
